+++ src/main/java/com/github/thencuber/intellijlinterplugin/inspection/AnnotatorService.java
@@ -1,12 +1,15 @@
 package com.github.thencuber.intellijlinterplugin.inspection;
 
+import com.github.thencuber.intellijlinterplugin.notifications.PluginNotification;
 import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer;
+import com.intellij.notification.Notification;
 import com.intellij.openapi.components.Service;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.openapi.vfs.VirtualFileManager;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
+import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
 import javax.xml.parsers.DocumentBuilder;
@@ -35,17 +38,27 @@ public final class AnnotatorService {
     public void updateAnnotators() {
         // Read config if possible
         String basePath = MyProject.getBasePath();
+        if(basePath == null) {
+            return;
+        }
         Path configPath = Paths.get(basePath, ".idea/linter.xml");
         if (Files.exists(configPath)) {
             System.out.println("Updating annotators using config file: " + configPath);
             VirtualFile configFile = VirtualFileManager.getInstance().findFileByNioPath(configPath);
+            if(configFile == null) {
+                return;
+            }
             try {
-                // DEBUG: Print full config file
-                // System.out.println(VfsUtil.loadText(configFile));
-
                 DocumentBuilderFactory builderFactory = DocumentBuilderFactory.newInstance();
                 DocumentBuilder builder = builderFactory.newDocumentBuilder();
 
+                try {
+                    builder.parse(configFile.getInputStream());
+                } catch(Exception e) {
+                    Notification failMessage = new PluginNotification(false, "Linter.xml is not a valid xml document.");
+                    failMessage.notify(MyProject);
+                    return;
+                }
                 Document document = builder.parse(configFile.getInputStream());
                 document.getDocumentElement().normalize();
                 ArrayList<AbstractAnnotator> Annotators = new ArrayList<>();
@@ -58,62 +71,112 @@ public final class AnnotatorService {
                     String currentMatcherType = currentMatcher.getAttribute("type");
                     switch (currentMatcherType) {
                         case "literal":
-                            String currentMatcherContent = currentMatcher.getElementsByTagName("literal").item(0).getTextContent();
-                            Annotators.add(new LiteralAnnotator(currentNote,currentSeverity,currentMatcherContent));
+                            Annotators.add(composeLiteralAnnotator(currentNote, currentSeverity, currentMatcher.getElementsByTagName("literal").item(0)));
                             break;
                         case "variable":
-                            String currentVariableName = null;
-                            if(!(currentMatcher.getElementsByTagName("variable").item(0).getTextContent() == null || Objects.equals(currentMatcher.getElementsByTagName("variable").item(0).getTextContent(), ""))) {
-                                currentVariableName = currentMatcher.getElementsByTagName("variable").item(0).getTextContent();
-                            }
-                            String currentVariableType = null;
-                            if(((Element) currentMatcher.getElementsByTagName("variable").item(0)).hasAttribute("type")) {
-                                currentVariableType = ((Element) currentMatcher.getElementsByTagName("variable").item(0)).getAttribute("type");
-                            }
-                            Annotators.add(new VariableAnnotator(currentNote,currentSeverity,currentVariableName,currentVariableType));
+                            Annotators.add(composeVariableAnnotator(currentNote, currentSeverity, currentMatcher.getElementsByTagName("variable").item(0)));
                             break;
                         case "method":
-                            String currentMethodReturnType = null;
-                            if(((Element) currentMatcher.getElementsByTagName("method").item(0)).hasAttribute("returnType")) {
-                                currentMethodReturnType = ((Element) currentMatcher.getElementsByTagName("method").item(0)).getAttribute("returnType");
-                            }
-                            String currentMethodName = null;
-                            if(((Element) currentMatcher.getElementsByTagName("method").item(0)).hasAttribute("name")) {
-                                currentMethodName = ((Element) currentMatcher.getElementsByTagName("method").item(0)).getAttribute("name");
-                            }
-                            ArrayList<String> currentMethodModifierList = null;
-                            if(((Element) currentMatcher.getElementsByTagName("method").item(0)).getElementsByTagName("modifier").getLength() > 0) {
-                                currentMethodModifierList = new ArrayList<>();
-                                NodeList XMLModifierList = ((Element) currentMatcher.getElementsByTagName("method").item(0)).getElementsByTagName("modifier");
-                                for (int j = 0; j < XMLModifierList.getLength(); j++) {
-                                    if(!(Objects.equals(XMLModifierList.item(j).getTextContent(),""))) {
-                                        currentMethodModifierList.add(XMLModifierList.item(j).getTextContent());
-                                    }
-                                }
-                            }
-                            ArrayList<String> currentMethodParameterList = null;
-                            if(((Element) currentMatcher.getElementsByTagName("method").item(0)).getElementsByTagName("parameter").getLength() > 0) {
-                                currentMethodParameterList = new ArrayList<>();
-                                NodeList XMLParameterList = ((Element) currentMatcher.getElementsByTagName("method").item(0)).getElementsByTagName("parameter");
-                                for (int j = 0; j < XMLParameterList.getLength(); j++) {
-                                    if(((Element) XMLParameterList.item(j)).hasAttribute("type")) {
-                                        currentMethodParameterList.add(((Element) XMLParameterList.item(j)).getAttribute("type"));
-                                    }
-                                }
-                            }
-                            Annotators.add(new MethodAnnotator(currentNote,currentSeverity,currentMethodName,currentMethodReturnType,currentMethodModifierList,currentMethodParameterList));
+                            Annotators.add(composeMethodAnnotator(currentNote, currentSeverity, currentMatcher.getElementsByTagName("method").item(0)));
                             break;
+                        case "class":
+                            Annotators.add(composeClassAnnotator(currentNote, currentSeverity, currentMatcher.getElementsByTagName("class").item(0)));
                         default: break;
                     }
 
                 }
                 AnnotatorList = Annotators;
                 DaemonCodeAnalyzer.getInstance(MyProject).restart();
+                Notification successMessage = new PluginNotification(true);
+                successMessage.notify(MyProject);
             } catch (Exception e) {
-                e.printStackTrace();
+                Notification failMessage = new PluginNotification(false, "Linter.xml is not following the supported structure.");
+                failMessage.notify(MyProject);
             }
         } else {
             System.out.println("Did not find linter.xml config file at " + configPath);
+            AnnotatorList = new ArrayList<>();
+            Notification failMessage = new PluginNotification(false, "No linter.xml file was found at " + configPath + ". Please create it, for the plugin to work.");
+            failMessage.notify(MyProject);
         }
+    }
+
+    private LiteralAnnotator composeLiteralAnnotator(String Note, String Severity, Node Matcher) {
+        String currentMatcherContent = Matcher.getTextContent();
+        return new LiteralAnnotator(Note,Severity,currentMatcherContent);
+    }
+
+    private VariableAnnotator composeVariableAnnotator(String Note, String Severity, Node Matcher) {
+        String currentVariableName = null;
+        if(!(Matcher.getTextContent() == null || Objects.equals(Matcher.getTextContent(), ""))) {
+            currentVariableName = Matcher.getTextContent();
+        }
+        String currentVariableType = null;
+        if(((Element) Matcher).hasAttribute("type")) {
+            currentVariableType = ((Element) Matcher).getAttribute("type");
+        }
+        return new VariableAnnotator(Note,Severity,currentVariableName,currentVariableType);
+    }
+
+    private MethodAnnotator composeMethodAnnotator(String Note, String Severity, Node Matcher) {
+        String currentMethodReturnType = null;
+        if(((Element) Matcher).hasAttribute("returnType")) {
+            currentMethodReturnType = ((Element) Matcher).getAttribute("returnType");
+        }
+        String currentMethodName = null;
+        if(((Element) Matcher).hasAttribute("name")) {
+            currentMethodName = ((Element) Matcher).getAttribute("name");
+        }
+        ArrayList<String> currentMethodModifierList = null;
+        if(((Element) Matcher).getElementsByTagName("modifier").getLength() > 0) {
+            currentMethodModifierList = new ArrayList<>();
+            NodeList XMLModifierList = ((Element) Matcher).getElementsByTagName("modifier");
+            for (int j = 0; j < XMLModifierList.getLength(); j++) {
+                if(!(Objects.equals(XMLModifierList.item(j).getTextContent(),""))) {
+                    currentMethodModifierList.add(XMLModifierList.item(j).getTextContent());
+                }
+            }
+        }
+        ArrayList<String> currentMethodParameterList = null;
+        if(((Element) Matcher).getElementsByTagName("parameter").getLength() > 0) {
+            currentMethodParameterList = new ArrayList<>();
+            NodeList XMLParameterList = ((Element) Matcher).getElementsByTagName("parameter");
+            for (int j = 0; j < XMLParameterList.getLength(); j++) {
+                if(((Element) XMLParameterList.item(j)).hasAttribute("type")) {
+                    currentMethodParameterList.add(((Element) XMLParameterList.item(j)).getAttribute("type"));
+                }
+            }
+        }
+        return new MethodAnnotator(Note,Severity,currentMethodName,currentMethodReturnType,currentMethodModifierList,currentMethodParameterList);
+    }
+
+    private ClassAnnotator composeClassAnnotator(String Note, String Severity, Node Matcher) {
+        String currentClassName = null;
+        if(((Element) Matcher).hasAttribute("name")) {
+            currentClassName = ((Element) Matcher).getAttribute("name");
+        }
+        String currentSuperClassName = null;
+        if(((Element) Matcher).hasAttribute("superClass")) {
+            currentSuperClassName = ((Element) Matcher).getAttribute("superClass");
+        }
+        ArrayList<String> currentClassInterfacesList = null;
+        if(((Element) Matcher).getElementsByTagName("interface").getLength() > 0) {
+            currentClassInterfacesList = new ArrayList<>();
+            NodeList XMLInterfacesList = ((Element) Matcher).getElementsByTagName("interface");
+            for (int j = 0; j < XMLInterfacesList.getLength(); j++) {
+                if(((Element) XMLInterfacesList.item(j)).hasAttribute("name")) {
+                    currentClassInterfacesList.add(((Element) XMLInterfacesList.item(j)).getAttribute("name"));
+                }
+            }
+        }
+        ArrayList<MethodAnnotator> currentClassMethodsList = null;
+        if(((Element) Matcher).getElementsByTagName("method").getLength() > 0) {
+            currentClassMethodsList = new ArrayList<>();
+            NodeList XMLMethodsList = ((Element) Matcher).getElementsByTagName("method");
+            for (int j = 0; j < XMLMethodsList.getLength(); j++) {
+                currentClassMethodsList.add(composeMethodAnnotator("INNER ANNOTATOR - NOTE IGNORED", "INFORMATION", XMLMethodsList.item(j)));
+            }
+        }
+        return new ClassAnnotator(Note,Severity,currentClassName,currentSuperClassName,currentClassInterfacesList,currentClassMethodsList);
     }
 }