+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support class-syntax in linter.xml config file 
+- Added success and fail notifications for updating linting rules
+
+### Changed
+- Save linter.xml on pressing UpdateButton in toolbar
+- Improved null-safety on several occasions
 
 ## [1.0.0]
 ### Added